Bio fixes
Now that iIcan see them all in one place, may as well fix glaring issues. Also added a field to Tate's bio to set precedent for DX-exclusive extra info. Also I do want to start using scare tag {sc} but only sparingly and on a low setting (such as when a character is about to cry).