Pushed fix to bios spacing
I hate it, though.

I took off the vbox, as for some reason it predicts the positioning of the largest child (out of all the bios) then centers every bios based on that.

I don't know why it doesn't just take an yanchor of the bios text and adhere it to the spacing policy.