Pushing better change that's less disgusting
So. After thinking about it. I realized what was happening.

yfill predicts the entirety of the vbox, so that causes the weird spacing in the text. So, I disabled yfill for that vbox, treated the vbox as its own sort of container for the text. Things get aligned nicely.

If we want to adjust the space between the bio and the quote, change the 'spacing' value in the vbox