add test video for perfect tate fight
"Tate2.webm" is too large to commit currently but it does exist. we'll need to compress it first.

there are still a few more elements i want to add but these ARE in a usable state. this was really hard to put together, but the vision is coming to life!

also added green screen background for all your screenshot needs.

or, you know, in case you also wanna f*ck around with in-game elements in a video