Jukebox revamp phase 2
Only reading songs that are actually unlocked now.

CSBI is updated to use the new syntax. The goal is to only need to use the same ID that you'd use to play the song in the first place and not have to memorize all the song titles/artists or worry about breaking the game if either value has a - in it like before. However, this is still VERY BROKEN and DOES NOT WORK YET