item collection + more. SEE NOTES
Please use collect() for this purpose. It'll tell you in console if you mess up somehow.

All existing items in the database have been made actually collectable... I think.

Also replaced the water bottle used in Christmas with a rarer and more festive variant (mostly to clear up a variable name conflict)

IMPORTANT: it appears that every time we try to play a movie in the background of a scene, it's broken. we need to fix this ASAP.

ALSO IMPORTANT: warning screens and CS's save data have been edited to be compatible with the new jukebox format. Also added a few songs he DID get, but weren't registered in jukebox at the time.

Somebody please find out if he got the interrogation ending, or if we missed any other routes he took, so we can add that content.