Fixed the minigame buttons!
For posterity, here's how it was done. The Replay() screen action takes an optional argument `locked`. If it's None (default), it *only* lets you jump there if you've seen the label before. If it's True, it doesn't do anything. If it's False, it always lets you jump to the label, regardless of if you've seen it before. Of course, we would *never* have seen these labels before, since they're bespoke labels designed only for the minigame menu.

Anyway, solved.