Removed something that's not ready yet
I'm sorry I did not mean to push this, Please do not read this file as I have made a mistake and didn't mean to send this. It would break my heart if anyone read this before it's finished. I hope everyone who pulls reads this first before checking the files